refactor(channel-whatsapp-cloud): collapse sidecar into plugin (H1, L1, M5)
H1 — Replace ProtectedPluginSurface sidecar cast with public plugin methods.

The webhook handler used to call `emitMessageReceivedFromMeta`,
`emitStatusEvent`, and `emitTemplateStatusChanged` from
`plugin-inbound-helpers.ts`, which reached into BaseChannelPlugin's
`protected` emit helpers via a `plugin as unknown as ProtectedPluginSurface`
cast. The cast silently breaks if BaseChannelPlugin ever renames any of
the seven `emit*` methods or the `eventBus` property — and the cast itself
violates the encapsulation contract every other channel plugin respects.

New surface on WhatsAppCloudPlugin (mirrors Gupshup's pattern):
- public `handleInboundMessage(instanceId, msg, contacts, dedupeCache)`
- public `handleStatusUpdate(instanceId, status)` — `sent` is a no-op
  (already emitted by sendMessage, see C2)
- public `handleTemplateStatusChanged(instanceId, update, extras?)` —
  publishes `template.status_changed` via `this.eventBus.publish` directly

The module-level `extractInboundContent` + `mapTemplateEventToStatus`
helpers move to the bottom of plugin.ts (still pure, no class state).

`handlers/webhook.ts` is now a thin dispatcher: signature verify → parse →
`plugin.handleInboundMessage(…)` / `plugin.handleStatusUpdate(…)` /
`handleTemplateStatusUpdate(db, update, plugin)`.

`plugin-inbound-helpers.ts` deleted.

M5 — Fix template fan-out across all connected instances.

The old webhook handler called `plugin.getConnectedInstanceIds()` and
fanned out `template.status_changed` once per connected instance,
regardless of which WABA owned the template. N instances sharing the
same Meta App would receive N duplicate events for a single template
lifecycle change.

`templates.ts::handleTemplateStatusUpdate` now resolves matching rows in
the local `whatsapp_templates` table via `metaId` OR `(name, language)`,
updates each row's status + rejectionReason, and notifies the plugin
once per matched row with the correct `(instanceId, templateId,
previousStatus)`. The webhook handler delegates to it (no more
`getConnectedInstanceIds` reliance), so the fan-out path is gone.

The function signature changed from `(db, instanceId, eventBus, update)`
to `(db, update, plugin)` — the local row is the source of truth for
`instanceId`. Tests updated to match.

L1 — Move `HealthStatus` to top-level type imports.

Was: `type HealthStatus = import('@omni/channel-sdk').HealthStatus`
declared *after* the class body, which is unconventional and confuses
both readers and tsdoc generators. Now imported alongside the other
`@omni/channel-sdk` types at the top of plugin.ts.
